+++ rabbitmq-tutorials/src/main/java/org/springframework/amqp/tutorials/tut1/Tut1Config.java
@@ -13,22 +13,47 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.amqp.tutorials.tut5;
+package org.springframework.amqp.tutorials.tut1;
 
-import org.springframework.amqp.core.TopicExchange;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.tutorials.util.SpringAwareExecutorWrapper;
+import org.springframework.context.Lifecycle;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 /**
- * @author Gary Russell
+ * @author Gary Russell, Scott Deeg
  *
  */
+@Profile("tut1")
 @Configuration
-public class CommonConfig {
+public class Tut1Config {
 
 	@Bean
-	public TopicExchange topic() {
-		return new TopicExchange("tut.topic");
+	public Queue hello() {
+		return new Queue("tut.hello");
+	}
+	
+	@Profile("receiver")
+	@Bean
+	public Tut1Receiver receiver() {
+		return new Tut1Receiver();
 	}
 
+	@Profile("sender")
+	@Configuration
+	public static class SenderConfig
+	{
+		@Bean
+		public Lifecycle wrappedSender(Tut1Sender sender) {
+			return new SpringAwareExecutorWrapper(sender);
+		}
+		
+		@Bean
+		public Tut1Sender sender()
+		{
+			return new Tut1Sender();
+		}
+	}
 }
