+++ rabbitmq-tutorials/src/main/java/org/springframework/amqp/tutorials/tut2/Tut2Config.java
@@ -16,19 +16,51 @@
 package org.springframework.amqp.tutorials.tut2;
 
 import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.tutorials.util.SpringAwareExecutorWrapper;
+import org.springframework.context.Lifecycle;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 
 /**
- * @author Gary Russell
+ * @author Gary Russell, Scott Deeg
  *
  */
+@Profile("tut2")
 @Configuration
-public class CommonConfig {
-
+public class Tut2Config {
+	
 	@Bean
 	public Queue hello() {
 		return new Queue("tut.hello");
 	}
 
+	@Profile("receiver")
+	public static class ReceiverConfig {
+		
+		@Bean
+		public Tut2Receiver receiver1() {
+			return new Tut2Receiver(1);
+		}
+
+		@Bean
+		public Tut2Receiver receiver2() {
+			return new Tut2Receiver(2);
+		}
+	}
+
+	@Profile("sender")
+	@Configuration
+	public static class SenderConfig
+	{
+		@Bean
+		public Lifecycle wrappedSender(Tut2Sender sender) {
+			return new SpringAwareExecutorWrapper(sender);
+		}
+		
+		@Bean
+		public Tut2Sender sender() {
+			return new Tut2Sender();
+		}
+	}
 }
