+++ rabbitmq-tutorials/src/main/java/org/springframework/amqp/tutorials/tut2/Tut2Receiver.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.amqp.tutorials.tut2;
+
+import org.apache.log4j.Logger;
+import org.springframework.amqp.rabbit.annotation.RabbitHandler;
+import org.springframework.amqp.rabbit.annotation.RabbitListener;
+import org.springframework.util.StopWatch;
+
+/**
+ *
+ * @author Gary Russell, Scott Deeg
+ *
+ */
+@RabbitListener(queues = "tut.hello")
+public class Tut2Receiver {
+
+	private static Logger logger = Logger.getLogger(Tut2Receiver.class);
+
+	private final int instance;
+
+	public Tut2Receiver(int i) {
+		this.instance = i;
+	}
+
+	@RabbitHandler
+	public void receive(String in) throws InterruptedException {
+		StopWatch watch = new StopWatch();
+		watch.start();
+		logger.error("instance " + this.instance + " [x] Received '" + in + "'");
+		dowork(in);
+		watch.stop();
+		logger.error("instance " + this.instance + " [x] Done in " + watch.getTotalTimeSeconds() + "s");
+	}
+
+	private void dowork(String in) throws InterruptedException {
+		for (char ch : in.toCharArray()) {
+			if (ch == '.') {
+				Thread.sleep(1000);
+			}
+		}
+	}
+}
