+++ rabbitmq-tutorials/src/main/java/org/springframework/amqp/tutorials/tut6/Tut6Config.java
@@ -1,0 +1,74 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.amqp.tutorials.tut6;
+
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.DirectExchange;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.tutorials.util.SpringAwareExecutorWrapper;
+import org.springframework.context.Lifecycle;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+/**
+ * @author Gary Russell
+ *
+ */
+@Profile("tut5")
+@Configuration
+public class Tut6Config {
+
+	@Profile("client")
+	public static class ClientConfig {
+		@Bean
+		public Lifecycle wrappedSender(Tut6Client client) {
+			return new SpringAwareExecutorWrapper(client);
+		}
+		
+		@Bean
+		public Tut6Client client() {
+	 	 	return new Tut6Client();
+		}
+	}
+
+	@Profile("server")
+	@Configuration
+	public static class ServerConfig
+	{
+		@Bean
+		public Queue queue() {
+			return new Queue("tut.rpc.requests");
+		}
+
+		@Bean
+		public DirectExchange exchange() {
+			return new DirectExchange("tut.rpc");
+		}
+
+		@Bean
+		public Binding binding() {
+			return BindingBuilder.bind(queue()).to(exchange()).with("rpc");
+		}
+
+		@Bean
+		public Tut6Server server() {
+			return new Tut6Server();
+		}
+	}
+
+}
