+++ rabbitmq-tutorials/src/main/resources/application-remote.yml
@@ -1,0 +1,5 @@
+spring:
+  rabbitmq:
+    host: rabbitserver
+    username: tutorial
+    password: tutorial
