+++ rabbitmq-tutorials/src/main/resources/application.yml
@@ -1,0 +1,8 @@
+#Two profiles are needed for the app:  role (sender|receiver) and logic
+spring:
+  profiles:
+    active: usage_message
+
+tutorial:
+  client:
+    duration: 10000
