+++ rabbitmq-tutorials/src/main/resources/banner.txt
@@ -1,0 +1,4 @@
+ __                 __   ___                  
+|__)_ |_ |_ .|_|\/|/  \   |    |_ _  _. _ | _ 
+| \(_||_)|_)||_|  |\_\/   | |_||_(_)| |(_||_) 
+
