+++ rabbitmq-tutorials/src/main/resources/help.txt
@@ -1,0 +1,32 @@
+RabbitMQ Tutorial Sample Application
+
+This application implements the class RabbitMQ tutorials in Spring AMQP.  It is a
+CLI app that uses Spring Profiles to control its behavior.
+
+Sample usage:
+java -jar rabbitmq-tutorials.jar --spring.profiles.active=tut1,sender
+or
+java -jar rabbitmq-tutorials.jar --spring.profiles.active=tut2,receiver
+
+The following tutorials are available:
+
+tut1,{sender|receiver}   Hello World
+tut2,{sender|receiver}   Work Queues
+tut3,{sender|receiver}   Work Queues
+tut4,{sender|receiver}   Work Queues
+tut5,{sender|receiver}   Work Queues
+tut6,{client|server}   Work Queues
+
+(The order of the profiles is not relevant.)
+
+Remote Hosts:
+
+To use to a remote RabbitMQ installation set the following properties as in this sample:
+
+spring.rabbitmq.host=your-rabbit-server
+spring.rabbitmq.username=tutorial-user
+spring.rabbitmq.password=tutorial-password
+
+To use this at runtime create a file called application-remote.yml (or properties) and set the properties in there.  Then use the 
+remote profile.  See the Spring Boot and Spring AMQP documentation for more information on setting application 
+properties and AMQP properties specifically.
